LPS-145823 Migrate FDS displays in `commerce-order-content-web`
