+++ modules/apps/commerce/commerce-order-content-web/build.gradle
@@ -21,6 +21,8 @@ dependencies {
 	compileOnly project(":apps:commerce:headless:headless-commerce:headless-commerce-delivery-cart-api")
 	compileOnly project(":apps:document-library:document-library-api")
 	compileOnly project(":apps:dynamic-data-mapping:dynamic-data-mapping-taglib")
+	compileOnly project(":apps:frontend-data-set:frontend-data-set-api")
+	compileOnly project(":apps:frontend-data-set:frontend-data-set-taglib")
 	compileOnly project(":apps:frontend-taglib:frontend-taglib")
 	compileOnly project(":apps:frontend-taglib:frontend-taglib-clay")
 	compileOnly project(":apps:frontend-taglib:frontend-taglib-soy")
