--- modules/apps/commerce/commerce-order-content-web/src/main/java/com/liferay/commerce/order/content/web/internal/frontend/constants/CommerceOrderDataSetConstants.java
@@ -1,45 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.commerce.order.content.web.internal.frontend.constants;
-
-/**
- * @author Marco Leo
- * @author Alessio Antonio Rendina
- */
-public class CommerceOrderDataSetConstants {
-
-	public static final String COMMERCE_DATA_SET_KEY_COMMERCE_WISH_LISTS =
-		"commerceWishLists";
-
-	public static final String COMMERCE_DATA_SET_KEY_IMPORT_ORDERS =
-		"commerceImportOrders";
-
-	public static final String COMMERCE_DATA_SET_KEY_PENDING_ORDER_ITEMS =
-		"commercePendingOrderItems";
-
-	public static final String COMMERCE_DATA_SET_KEY_PENDING_ORDERS =
-		"commercePendingOrders";
-
-	public static final String COMMERCE_DATA_SET_KEY_PLACED_ORDER_ITEMS =
-		"commercePlacedOrderItems";
-
-	public static final String COMMERCE_DATA_SET_KEY_PLACED_ORDERS =
-		"commercePlacedOrders";
-
-	public static final String
-		COMMERCE_DATA_SET_KEY_PREVIEW_COMMERCE_ORDER_ITEMS =
-			"previewCommerceOrderItems";
-
-}
