+++ modules/apps/commerce/commerce-order-content-web/src/main/java/com/liferay/commerce/order/content/web/internal/frontend/data/set/provider/CommerceWishListFDSActionProvider.java
@@ -1,0 +1,57 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.commerce.order.content.web.internal.frontend.data.set.provider;
+
+import com.liferay.commerce.order.content.web.internal.constants.CommerceOrderFDSNames;
+import com.liferay.commerce.order.content.web.internal.frontend.data.set.util.CommerceOrderFDSUtil;
+import com.liferay.commerce.order.content.web.internal.model.WishList;
+import com.liferay.frontend.data.set.provider.FDSActionProvider;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemBuilder;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.language.LanguageUtil;
+
+import java.util.Collections;
+import java.util.List;
+
+import javax.servlet.http.HttpServletRequest;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Alessio Antonio Rendina
+ */
+@Component(
+	enabled = false, immediate = true,
+	property = "fds.data.provider.key=" + CommerceOrderFDSNames.WISH_LISTS,
+	service = FDSActionProvider.class
+)
+public class CommerceWishListFDSActionProvider implements FDSActionProvider {
+
+	@Override
+	public List<DropdownItem> getDropdownItems(
+			long groupId, HttpServletRequest httpServletRequest, Object model)
+		throws PortalException {
+
+		return Collections.singletonList(
+			DropdownItemBuilder.setHref(
+				CommerceOrderFDSUtil.getWishListCommerceOrderPreviewURL(
+					(WishList)model, httpServletRequest)
+			).setLabel(
+				LanguageUtil.get(httpServletRequest, "select")
+			).build());
+	}
+
+}
