+++ modules/apps/commerce/commerce-order-content-web/src/main/java/com/liferay/commerce/order/content/web/internal/frontend/data/set/provider/PendingCommerceOrderItemFDSActionProvider.java
@@ -1,0 +1,129 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.commerce.order.content.web.internal.frontend.data.set.provider;
+
+import com.liferay.commerce.model.CommerceOrder;
+import com.liferay.commerce.order.content.web.internal.constants.CommerceOrderFDSNames;
+import com.liferay.commerce.order.content.web.internal.model.OrderItem;
+import com.liferay.commerce.product.model.CPInstance;
+import com.liferay.commerce.product.service.CPInstanceLocalService;
+import com.liferay.commerce.product.util.CPDefinitionHelper;
+import com.liferay.commerce.service.CommerceOrderService;
+import com.liferay.frontend.data.set.provider.FDSActionProvider;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemListBuilder;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.security.permission.ActionKeys;
+import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
+import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.WebKeys;
+
+import java.util.Collections;
+import java.util.List;
+
+import javax.servlet.http.HttpServletRequest;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Alessio Antonio Rendina
+ */
+@Component(
+	enabled = false, immediate = true,
+	property = "fds.data.provider.key=" + CommerceOrderFDSNames.PENDING_ORDER_ITEMS,
+	service = FDSActionProvider.class
+)
+public class PendingCommerceOrderItemFDSActionProvider
+	implements FDSActionProvider {
+
+	@Override
+	public List<DropdownItem> getDropdownItems(
+			long groupId, HttpServletRequest httpServletRequest, Object model)
+		throws PortalException {
+
+		OrderItem orderItem = (OrderItem)model;
+
+		if (orderItem.getParentOrderItemId() > 0) {
+			return Collections.emptyList();
+		}
+
+		CommerceOrder commerceOrder = _commerceOrderService.getCommerceOrder(
+			orderItem.getOrderId());
+
+		return DropdownItemListBuilder.add(
+			() -> _modelResourcePermission.contains(
+				PermissionThreadLocal.getPermissionChecker(), commerceOrder,
+				ActionKeys.VIEW),
+			dropdownItem -> {
+				CPInstance cpInstance = _cpInstanceLocalService.fetchCPInstance(
+					orderItem.getCPInstanceId());
+
+				if (cpInstance == null) {
+					dropdownItem.setHref(StringPool.BLANK);
+				}
+				else {
+					ThemeDisplay themeDisplay =
+						(ThemeDisplay)httpServletRequest.getAttribute(
+							WebKeys.THEME_DISPLAY);
+
+					dropdownItem.setHref(
+						_cpDefinitionHelper.getFriendlyURL(
+							cpInstance.getCPDefinitionId(), themeDisplay));
+				}
+
+				dropdownItem.setLabel(
+					LanguageUtil.get(httpServletRequest, "view"));
+			}
+		).add(
+			() ->
+				_modelResourcePermission.contains(
+					PermissionThreadLocal.getPermissionChecker(), commerceOrder,
+					ActionKeys.UPDATE) &&
+				commerceOrder.isOpen(),
+			dropdownItem -> {
+				dropdownItem.putData("method", "delete");
+				dropdownItem.setHref(
+					_getDeleteCommerceOrderItemURL(orderItem.getOrderItemId()));
+				dropdownItem.setLabel(
+					LanguageUtil.get(httpServletRequest, "delete"));
+				dropdownItem.setTarget("async");
+			}
+		).build();
+	}
+
+	private String _getDeleteCommerceOrderItemURL(long commerceOrderItemId) {
+		return "/o/headless-commerce-delivery-cart/v1.0/cart-items/" +
+			commerceOrderItemId;
+	}
+
+	@Reference
+	private CommerceOrderService _commerceOrderService;
+
+	@Reference
+	private CPDefinitionHelper _cpDefinitionHelper;
+
+	@Reference
+	private CPInstanceLocalService _cpInstanceLocalService;
+
+	@Reference(
+		target = "(model.class.name=com.liferay.commerce.model.CommerceOrder)"
+	)
+	private ModelResourcePermission<CommerceOrder> _modelResourcePermission;
+
+}
