--- modules/apps/commerce/commerce-order-content-web/src/main/java/com/liferay/commerce/order/content/web/internal/frontend/taglib/clay/data/set/provider/ImportCommerceOrderDataSetDataProvider.java
@@ -1,132 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.commerce.order.content.web.internal.frontend.taglib.clay.data.set.provider;
-
-import com.liferay.commerce.model.CommerceOrder;
-import com.liferay.commerce.order.content.web.internal.frontend.constants.CommerceOrderDataSetConstants;
-import com.liferay.commerce.order.content.web.internal.frontend.util.CommerceOrderClayTableUtil;
-import com.liferay.commerce.order.content.web.internal.model.Order;
-import com.liferay.commerce.product.model.CommerceChannel;
-import com.liferay.commerce.product.service.CommerceChannelLocalService;
-import com.liferay.commerce.service.CommerceOrderService;
-import com.liferay.frontend.taglib.clay.data.Filter;
-import com.liferay.frontend.taglib.clay.data.Pagination;
-import com.liferay.frontend.taglib.clay.data.set.provider.ClayDataSetDataProvider;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.search.Sort;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.util.Collections;
-import java.util.List;
-
-import javax.servlet.http.HttpServletRequest;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Alessio Antonio Rendina
- */
-@Component(
-	enabled = false, immediate = true,
-	property = "clay.data.provider.key=" + CommerceOrderDataSetConstants.COMMERCE_DATA_SET_KEY_IMPORT_ORDERS,
-	service = ClayDataSetDataProvider.class
-)
-public class ImportCommerceOrderDataSetDataProvider
-	implements ClayDataSetDataProvider<Order> {
-
-	@Override
-	public List<Order> getItems(
-			HttpServletRequest httpServletRequest, Filter filter,
-			Pagination pagination, Sort sort)
-		throws PortalException {
-
-		ThemeDisplay themeDisplay =
-			(ThemeDisplay)httpServletRequest.getAttribute(
-				WebKeys.THEME_DISPLAY);
-
-		CommerceChannel commerceChannel =
-			_commerceChannelLocalService.fetchCommerceChannelBySiteGroupId(
-				themeDisplay.getScopeGroupId());
-
-		if (commerceChannel == null) {
-			return Collections.emptyList();
-		}
-
-		List<CommerceOrder> commerceOrders =
-			_commerceOrderService.getUserCommerceOrders(
-				commerceChannel.getCompanyId(), commerceChannel.getGroupId(),
-				filter.getKeywords(), pagination.getStartPosition(),
-				pagination.getEndPosition());
-
-		long commerceOrderId = ParamUtil.getLong(
-			httpServletRequest, "commerceOrderId");
-
-		CommerceOrder commerceOrder = _commerceOrderService.fetchCommerceOrder(
-			commerceOrderId);
-
-		if (commerceOrder != null) {
-			commerceOrders.remove(commerceOrder);
-		}
-
-		return CommerceOrderClayTableUtil.getOrders(
-			commerceOrders, null, null, commerceChannel.getPriceDisplayType(),
-			true, themeDisplay);
-	}
-
-	@Override
-	public int getItemsCount(
-			HttpServletRequest httpServletRequest, Filter filter)
-		throws PortalException {
-
-		ThemeDisplay themeDisplay =
-			(ThemeDisplay)httpServletRequest.getAttribute(
-				WebKeys.THEME_DISPLAY);
-
-		CommerceChannel commerceChannel =
-			_commerceChannelLocalService.fetchCommerceChannelBySiteGroupId(
-				themeDisplay.getScopeGroupId());
-
-		if (commerceChannel == null) {
-			return 0;
-		}
-
-		int userCommerceOrdersCount =
-			(int)_commerceOrderService.getUserCommerceOrdersCount(
-				commerceChannel.getCompanyId(), commerceChannel.getGroupId(),
-				filter.getKeywords());
-
-		long commerceOrderId = ParamUtil.getLong(
-			httpServletRequest, "commerceOrderId");
-
-		CommerceOrder commerceOrder = _commerceOrderService.fetchCommerceOrder(
-			commerceOrderId);
-
-		if (commerceOrder != null) {
-			return userCommerceOrdersCount - 1;
-		}
-
-		return userCommerceOrdersCount;
-	}
-
-	@Reference
-	private CommerceChannelLocalService _commerceChannelLocalService;
-
-	@Reference
-	private CommerceOrderService _commerceOrderService;
-
-}
