+++ modules/apps/commerce/commerce-order-content-web/src/main/resources/META-INF/resources/pending_commerce_orders/b2b/view.jsp
@@ -33,13 +33,10 @@ CommerceAccount commerceAccount = commerceOrderContentDisplayContext.getCommerce
 	displayStyleGroupId="<%= commerceOrderContentDisplayContext.getDisplayStyleGroupId(CommercePortletKeys.COMMERCE_OPEN_ORDER_CONTENT) %>"
 	entries="<%= commerceOrderSearchContainer.getResults() %>"
 >
-	<clay:data-set-display
-		dataProviderKey="<%= CommerceOrderDataSetConstants.COMMERCE_DATA_SET_KEY_PENDING_ORDERS %>"
-		id="<%= CommerceOrderDataSetConstants.COMMERCE_DATA_SET_KEY_PENDING_ORDERS %>"
+	<frontend-data-set:classic-display
+		dataProviderKey="<%= CommerceOrderFDSNames.PENDING_ORDERS %>"
+		id="<%= CommerceOrderFDSNames.PENDING_ORDERS %>"
 		itemsPerPage="<%= 10 %>"
-		namespace="<%= liferayPortletResponse.getNamespace() %>"
-		pageNumber="<%= 1 %>"
-		portletURL="<%= commerceOrderContentDisplayContext.getPortletURL() %>"
 		style="stacked"
 	/>
