+++ modules/apps/commerce/commerce-order-content-web/src/main/resources/META-INF/resources/placed_commerce_orders/b2b/view.jsp
@@ -27,13 +27,10 @@
 	displayStyleGroupId="<%= commerceOrderContentDisplayContext.getDisplayStyleGroupId(CommercePortletKeys.COMMERCE_ORDER_CONTENT) %>"
 	entries="<%= commerceOrderSearchContainer.getResults() %>"
 >
-	<clay:data-set-display
-		dataProviderKey="<%= CommerceOrderDataSetConstants.COMMERCE_DATA_SET_KEY_PLACED_ORDERS %>"
-		id="<%= CommerceOrderDataSetConstants.COMMERCE_DATA_SET_KEY_PLACED_ORDERS %>"
+	<frontend-data-set:classic-display
+		dataProviderKey="<%= CommerceOrderFDSNames.PLACED_ORDERS %>"
+		id="<%= CommerceOrderFDSNames.PLACED_ORDERS %>"
 		itemsPerPage="<%= 10 %>"
-		namespace="<%= liferayPortletResponse.getNamespace() %>"
-		pageNumber="<%= 1 %>"
-		portletURL="<%= commerceOrderContentDisplayContext.getPortletURL() %>"
 		style="stacked"
 	/>
 </liferay-ddm:template-renderer>
