+++ modules/apps/commerce/commerce-order-content-web/src/main/resources/META-INF/resources/placed_commerce_orders/view_commerce_order_details.jsp
@@ -308,16 +308,13 @@ if (commerceOrder != null) {
 		contextParams.put("commerceOrderId", String.valueOf(commerceOrder.getCommerceOrderId()));
 		%>
 
-		<clay:data-set-display
+		<frontend-data-set:classic-display
 			contextParams="<%= contextParams %>"
-			dataProviderKey="<%= CommerceOrderDataSetConstants.COMMERCE_DATA_SET_KEY_PLACED_ORDER_ITEMS %>"
-			id="<%= CommerceOrderDataSetConstants.COMMERCE_DATA_SET_KEY_PLACED_ORDER_ITEMS %>"
+			dataProviderKey="<%= CommerceOrderFDSNames.PLACED_ORDER_ITEMS %>"
+			id="<%= CommerceOrderFDSNames.PLACED_ORDER_ITEMS %>"
 			itemsPerPage="<%= 10 %>"
-			namespace="<%= liferayPortletResponse.getNamespace() %>"
 			nestedItemsKey="orderItemId"
 			nestedItemsReferenceKey="orderItems"
-			pageNumber="<%= 1 %>"
-			portletURL="<%= commerceOrderContentDisplayContext.getPortletURL() %>"
 			style="stacked"
 		/>
 	</div>
